refactor: startup and terminate
